requirements.txt
py-eureka-client
config
flask
flask_cors
google-generativeai
python-dotenv
endpoints

